requirements.txt
absl-py==2.1.0
alembic==1.13.1
aniso8601==9.0.1
astunparse==1.6.3
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
docker==7.0.0
entrypoints==0.4
exceptiongroup==1.2.0
filelock==3.13.4
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.3.1
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
grpcio==1.62.1
h5py==3.11.0
idna==3.7
importlib_metadata==7.1.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.4.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==18.1.1
Mako==1.3.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
ml-dtypes==0.3.2
mlflow==2.11.3
mpmath==1.3.0
namex==0.0.7
networkx==3.3
numpy==1.26.4
nvidia-cuda-runtime-cu12==12.4.127
nvidia-pyindex==1.0.9
oauthlib==3.2.2
opt-einsum==3.3.0
optree==0.11.0
packaging==23.2
pandas==2.2.2
pillow==10.3.0
pluggy==1.4.0
protobuf==4.25.3
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.17.2
pyparsing==3.1.2
pytest==8.1.1
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
querystring-parser==1.2.4
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.29
sqlparse==0.4.4
sympy==1.12
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.4.0
threadpoolctl==3.4.0
tomli==2.0.1
torch==2.2.2+cu121
torchaudio==2.2.2+cu121
torchvision==0.17.2+cu121
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
waitress==3.0.0
Werkzeug==3.0.2
wrapt==1.16.0
zipp==3.18.1
